Rich Text Editor enabling

+++ app/Http/Controllers/ProjectController.php
@@ -5,12 +5,21 @@ namespace App\Http\Controllers;
 use App\Models\Project;
 use App\Http\Requests\StoreProjectRequest;
 use App\Http\Requests\UpdateProjectRequest;
+use App\Models\Course;
+use Gate;
 
 class ProjectController extends Controller
 {
 
+    
+
+
     public function index()
     {
+
+        if (Gate::inspect('viewAny', Project::class)->denied()) {
+            return redirect()->route('index.index')->with('error', '403 Unauthorized access: Contact your admin!');
+        }
 
         $projects = Project::query()
             ->latest('id')
@@ -19,14 +28,18 @@ class ProjectController extends Controller
 
         return inertia('Project/Index', [
             'projects' => $projects,
-            'filters' => [],
-            'user' => auth()->user(),
+            'filters' => []
         ]);
     }
 
 
     public function create()
     {
+
+        if (Gate::inspect('create', Project::class)->denied()) {
+            return redirect()->route('project.index')->with('error', '403 Unauthorized access: Contact your admin!');
+        }
+
         return inertia('Project/Create', [
             'user' => auth()->user(),
         ]);
@@ -38,12 +51,15 @@ class ProjectController extends Controller
     public function store(StoreProjectRequest $request)
     {
 
+        if (Gate::inspect('create', Project::class)->denied()) {
+            return redirect()->route('index.index')->with('error', '403 Unauthorized access: Contact your admin!');
+        }
 
         $validated = $request->validated();
 
         $project = Project::create([
             'title' => $validated['title'],
-            'created_by' => auth()->id(),
+            'created_by' => auth()->user()->id,
         ]);
 
         return redirect()->route('projects.index')->with('success', 'Project was created!');
@@ -52,16 +68,29 @@ class ProjectController extends Controller
     public function show(Project $project)
     {
 
+        if (Gate::inspect('view', $project)->denied()) {
+            return redirect()->route('index.index')->with('error', '403 Unauthorized access: Contact your admin!');
+        }
+
+        $courses = Course::query()
+            ->where('project_id', $project->id)
+            ->latest('id')
+            ->paginate(10);
 
         return inertia('Project/Show', [
             'project' => $project,
             'user' => auth()->user(),
+            'courses' => $courses,
         ]);
     }
 
 
     public function edit(Project $project)
     {
+
+        if (Gate::inspect('update', $project)->denied()) {
+            return redirect()->route('index.index')->with('error', '403 Unauthorized access: Contact your admin!');
+        }
 
 
         return inertia('Project/Edit', [
@@ -74,6 +103,9 @@ class ProjectController extends Controller
     public function update(UpdateProjectRequest $request, Project $project)
     {
 
+        if (Gate::inspect('update', $project)->denied()) {
+            return redirect()->route('index.index')->with('error', '403 Unauthorized access: Contact your admin!');
+        }
 
         $validated = $request->validated();
 
@@ -87,6 +119,9 @@ class ProjectController extends Controller
 
     public function destroy(Project $project)
     {
+        if (Gate::inspect('delete', $project)->denied()) {
+            return redirect()->route('index.index')->with('error', '403 Unauthorized access: Contact your admin!');
+        }
 
         $project->delete();
